rx: fix used tracks check in update path
When we update an RX stream, we need to check early whether the tracks
referenced by the new stream are available. We can only do that after giving
up the current allocation, and we need to restore that in case of any
errors.

This was previously only done when the number of channels changed, but we
need to do this in all cases.