go: refine RTP syncer
The algorithm to sync the local device to the PTP network requires a
quantization of the PTP time to a sample counter.

The current implementation does a simple integer division, which can lead
to rounding errors when the product of the PTP time and the sample rate is
close to a multiple of `nanoSecondsPerSecond`. This can lead to the local
device updating the RTP offset frequently, which causes audio artifacts.
Hence, the implementation allowed for a deviation of ±1 sample.

This, however, is not sufficient for applications requiring higher time
precision.

This patch introduces an improved method by resolving the sample counter to
sub-samples. The number of sub-samples per sample is currently set to 8,
and the allowed deviation is set to a quarter sample. This refined
measurement enables sample-accurate comparisons during periodic checks
while avoiding unnecessary updates caused by integer divisions.

Only if the offset changes by more than a quarter sample, the RTP offset
is adjusted in the FPGA.

Also add some comments and a test suite.