go: refine RTP syncer
The current implementation of the RTP syncer allows for a deviation
of ±1 sample, which is insufficient for applications requiring
higher time precision.

The original approach aimed to address rounding issues from integer
division when the PTP time coincided with the boundary of
`nanoSecondsPerSecond / sampleRate` and sought to minimize frequent
updates to the RTP offset in such cases.

This patch introduces an improved method by resolving the sample
counter to sub-samples, and the number of sub-samples per samples
is currently set to 4.

These refined measurements enable more accurate comparisons during
periodic checks. If the offset changes by half a sample or more,
the FPGA register is updated accordingly.

Also add some comments and a test suite.